requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pymongo>=4.11,<5.0
pydantic-settings>=2.6,<3.0
python-dotenv>=1.0,<2.0
dnspython>=2.6,<3.0
